+++ lib/cye/src/repair.cpp
@@ -16,9 +16,9 @@
 #include "cye/solution.hpp"
 
 struct CargoDPCell {
-  CargoDPCell() : dist(std::numeric_limits<float>::infinity()), prev(0), inserted(false) {}
+  CargoDPCell() : dist(std::numeric_limits<double>::infinity()), prev(0), inserted(false) {}
 
-  float dist;
+  double dist;
   unsigned prev;
   bool inserted;
 };
@@ -29,7 +29,7 @@ auto cye::patch_cargo_optimally(Solution &solution, unsigned bin_cnt) -> void {
   auto dp = std::vector(bin_cnt, std::vector(visited_node_cnt + 2, CargoDPCell()));
 
   // Amount of cargo per bin
-  auto cargo_quant = instance.cargo_capacity() / static_cast<float>(bin_cnt - 1);
+  auto cargo_quant = instance.cargo_capacity() / static_cast<double>(bin_cnt - 1);
 
   // Forward pass
 
@@ -78,7 +78,7 @@ auto cye::patch_cargo_optimally(Solution &solution, unsigned bin_cnt) -> void {
 
   // Find the smallest cost in the last column
   auto ind = 0u;
-  auto min_cost = std::numeric_limits<float>::infinity();
+  auto min_cost = std::numeric_limits<double>::infinity();
   for (auto i = 0u; i < bin_cnt; ++i) {
     if (dp[i][visited_node_cnt + 1].dist < min_cost) {
       min_cost = dp[i][visited_node_cnt + 1].dist;
@@ -122,10 +122,10 @@ auto cye::patch_cargo_trivially(Solution &solution) -> void {
   solution.add_patch(std::move(patch));
 }
 
-auto find_charging_station(const cye::Instance &instance, size_t node1_id, size_t node2_id, float remaining_battery)
+auto find_charging_station(const cye::Instance &instance, size_t node1_id, size_t node2_id, double remaining_battery)
     -> std::optional<size_t> {
   auto best_station_id = std::optional<size_t>{};
-  auto min_distance = std::numeric_limits<float>::infinity();
+  auto min_distance = std::numeric_limits<double>::infinity();
 
   if (node1_id != instance.depot_id() && node2_id != instance.depot_id() &&
       remaining_battery >= instance.energy_required(node1_id, instance.depot_id())) {
@@ -200,7 +200,7 @@ cye::OptimalEnergyRepair::OptimalEnergyRepair(std::shared_ptr<Instance> instance
 
 auto cye::OptimalEnergyRepair::compute_cs_dist_mat_() -> void {
   auto cs_cnt = instance_->charging_station_cnt() + 1;
-  cs_dist_mat_ = std::vector(cs_cnt, std::vector(cs_cnt, std::numeric_limits<float>::infinity()));
+  cs_dist_mat_ = std::vector(cs_cnt, std::vector(cs_cnt, std::numeric_limits<double>::infinity()));
 
   for (auto i = 0UZ; i < cs_cnt; ++i) {
     cs_dist_mat_[i][i] = 0.f;
@@ -224,7 +224,7 @@ auto cye::OptimalEnergyRepair::reset_() -> void {
 }
 
 auto cye::OptimalEnergyRepair::find_between_(size_t start_node_id, size_t goal_node_id)
-    -> std::optional<std::pair<std::vector<size_t>, float>> {
+    -> std::optional<std::pair<std::vector<size_t>, double>> {
   reset_();
 
   unvisited_queue_.emplace(start_node_id, start_node_id, 0.f, instance_->distance(start_node_id, goal_node_id));
@@ -284,7 +284,7 @@ auto cye::OptimalEnergyRepair::fill_dp(Solution &solution, unsigned bin_cnt) -> 
   auto dp = std::vector(bin_cnt, std::vector(visited_node_cnt, DPCell()));
 
   // Energy per bin
-  auto energy_per_bin = instance.battery_capacity() / static_cast<float>(bin_cnt - 1);
+  auto energy_per_bin = instance.battery_capacity() / static_cast<double>(bin_cnt - 1);
   auto cs_cnt = instance_->charging_station_cnt() + 1;
 
   // Forward pass
@@ -309,7 +309,7 @@ auto cye::OptimalEnergyRepair::fill_dp(Solution &solution, unsigned bin_cnt) -> 
 
         auto distance_to_entry_cs = instance_->distance(previous_node_id, entry_node_id);
         auto energy_to_entry_cs = distance_to_entry_cs * instance.energy_consumption();
-        auto remaining_battery = static_cast<float>(i) * energy_per_bin;
+        auto remaining_battery = static_cast<double>(i) * energy_per_bin;
 
         if (energy_to_entry_cs > remaining_battery) {
           continue;
@@ -369,7 +369,7 @@ auto cye::OptimalEnergyRepair::patch(Solution &solution, unsigned bin_cnt) -> vo
 
   // Find the smallest cost in the last column
   auto ind = 0u;
-  auto min_cost = std::numeric_limits<float>::infinity();
+  auto min_cost = std::numeric_limits<double>::infinity();
   for (auto i = 0u; i < bin_cnt; ++i) {
     if (dp[i][visited_node_cnt - 1].dist < min_cost) {
       min_cost = dp[i][visited_node_cnt - 1].dist;
@@ -377,7 +377,7 @@ auto cye::OptimalEnergyRepair::patch(Solution &solution, unsigned bin_cnt) -> vo
     }
   }
 
-  if (min_cost == std::numeric_limits<float>::infinity()) {
+  if (min_cost == std::numeric_limits<double>::infinity()) {
     throw std::runtime_error("Solution not found");
   }
